added useToken and refactoring
